- Implemented Get - Mappers - Web Api Filters for Exception and Validation including Filter Registration in WebAPiConfig - Trivial Logger - Implemented Put Post and Delete and renamed to Update and Add

+++ GeographicalLocationService/ExternalServices/Weather/Main.cs
@@ -8,10 +8,10 @@
 		public double Temp { get; set; }
 
 		[JsonProperty("pressure")]
-		public long Pressure { get; set; }
+		public int Pressure { get; set; }
 
 		[JsonProperty("humidity")]
-		public long Humidity { get; set; }
+		public int Humidity { get; set; }
 
 		[JsonProperty("temp_min")]
 		public double TempMin { get; set; }
