+++ GeographicalLocationService/Models/CurrentWeather.cs
@@ -6,12 +6,12 @@
 
 		public string Description { get; set; }
 
-		public decimal Temperature { get; set; }
+		public double Temperature { get; set; }
 
 		public int Pressure { get; set; }
 
 		public int Humidity { get; set; }
 
-		public decimal WindSpeed { get; set; }
+		public double WindSpeed { get; set; }
 	}
 }
